uploader/requirements.txt
APScheduler==3.4.0
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
dill==0.2.7.1
enum34==1.1.6
fake-useragent==0.1.8
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
google-api-core==0.1.1
google-auth==1.2.0
google-cloud-core==0.28.0
google-cloud-firestore==0.28.0
google-gax==0.15.15
googleapis-common-protos==1.5.3
grpcio==1.4.0
idna==2.6
ply==3.8
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pyreadline==2.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
twython==3.4.0
tzlocal==1.4
urllib3==1.22
